Correcion Decorador documento
