--- src/main/java/com/cenfotec/develectronicos/entities/Archivo.java
@@ -1,8 +1,0 @@
-package com.cenfotec.develectronicos.entities;
-
-public abstract class Archivo  {
-	
-    @Override
-    public abstract String toString();
-
-}
