+++ src/main/java/com/cenfotec/develectronicos/entities/Documento.java
@@ -1,60 +1,7 @@
 package com.cenfotec.develectronicos.entities;
 
-import java.time.LocalDateTime;
 
-public class Documento {
-	 private String idDoc;
-	 private String tipoTramite;
-	 private String responsable;
-	 private String productoId;
-	 private LocalDateTime fechaCreacion; 
-	 private int pasoActual;
- 
-	 public Documento(String idDoc, String tipoTramite, String responsable, String productId, 
-			 String stockStatus) {
-		this.idDoc= idDoc;
-		this.tipoTramite = tipoTramite;
-		this.responsable = responsable;
-		this.productoId = productId;
-		this.fechaCreacion = LocalDateTime.now();
-		this.pasoActual =1;
-	}
-	public Documento(String idDoc, String tipoTramite, String responsable, String productId, 
-			 String stockStatus,LocalDateTime fecha, int paso ) {
-		this.idDoc= idDoc;
-		this.tipoTramite = tipoTramite;
-		this.responsable = responsable;
-		this.productoId = productId;
-		this.fechaCreacion = LocalDateTime.now();
-		this.pasoActual = paso;
-	}
-
-	public int getPasoActual() {
-		return pasoActual;
-	}
+public abstract class  Documento {
 	
-	public void setPasoActual(int pasoActual) {
-		this.pasoActual = pasoActual;
-	}
-	
-	public String getIdDoc() {
-		return idDoc;
-	}
-	
-	public String getTipoTramite() {
-		return tipoTramite;
-	}
-
-	public String getProductoId() {
-		return productoId;
-	}
-
-	public String getResponsable() {
-		return responsable;
-	}
-	
-	public LocalDateTime getFechaCreacion() {
-		return fechaCreacion;
-	}
-
+	public abstract String toString();
 }
