+++ src/main/java/com/cenfotec/develectronicos/entities/extras/OrdenFacturacion.java
@@ -1,13 +1,13 @@
 package com.cenfotec.develectronicos.entities.extras;
 
-import com.cenfotec.develectronicos.entities.Documento;
+import com.cenfotec.develectronicos.entities.Orden;
 
 public class OrdenFacturacion extends DecoradorDocumento{
 
-	private Documento doc;
+	private Orden doc;
 	private String idCliente;
 	
-	public OrdenFacturacion(Documento doc, String idCliente) {
+	public OrdenFacturacion(Orden doc, String idCliente) {
 		this.doc = doc;
 		this.idCliente = idCliente;
 	}
@@ -19,7 +19,7 @@ public class OrdenFacturacion extends DecoradorDocumento{
 	@Override
 	public String toString() {
 		String mensaje;
-		mensaje= super.toString() + "\n"+
+		mensaje= doc.toString() + "\n"+
 				"Id del cliente"+ this.idCliente;		
 		return mensaje;
 	}
